% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{anaR-package}
\alias{anaR}
\alias{anaR-package}
\title{anaR: Acute Needs Analysis Engine}
\description{
Validates, flags, and exports ANA humanitarian data.
}
\author{
\strong{Maintainer}: gnoblet \email{gnoblet@zaclys.net}

}
\keyword{internal}
