% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdive.R
\name{dd_col_widths}
\alias{dd_col_widths}
\title{Column widths vector for a deep-dive sheet}
\usage{
dd_col_widths(n_hyps)
}
\arguments{
\item{n_hyps}{Integer number of hypothesis columns.}
}
\value{
Numeric vector of column widths.
}
\description{
Column widths vector for a deep-dive sheet
}
