% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdive.R
\name{dd_table_headers}
\alias{dd_table_headers}
\title{Table header row labels for a deep-dive sheet}
\usage{
dd_table_headers(hyp_ids)
}
\arguments{
\item{hyp_ids}{Character vector of hypothesis IDs (e.g. \code{c("H1","H2")}).}
}
\value{
Character vector of header labels.
}
\description{
Table header row labels for a deep-dive sheet
}
